Feature: pruning dead ends works with new cell struct

+++ src/maze_gen/dead_end.h
@@ -5,7 +5,7 @@
 #include <stdbool.h>
 
 bool is_dead_end(Cell* cell);
-uint8_t get_open_end(Cell* cell);
+int get_open_end(Cell* cell);
 int prune_dead_ends(Cell* cells, int rows, int cols);
 
 #endif
